requirements.txt
Flask==2.2.3
pyotp==2.6.0
firebase-admin==6.1.0
pyrebase4==4.5.0
pycryptodome==3.15.0
requests==2.28.1
flask-session==0.4.0
